requirements.txt
django>=4.2
whitenoise>=6.6
google-adk>=0.3.0
google-genai>=1.0.0
gunicorn>=21.2.0